Update barcodeMaker.py
Added a function to ask the user for the number of strings they want converted into barcodes. Those codes are stored in a folder on the desktop.

Excel sheet functionality is commented out for now, later on I'll add a function which copies the images into an excel sheet stored in the barcodes folder.