Update barcodeMaker.py
Added the excel sheet creation loop, doesn't quite work though. It adds thestrings and the first image to the sheet but the others don't get added for some reason. 

Also Excel keeps throwing an error saying there is something wrong with the worksheet. "We found an error with some content in myCustomBarcodes.xlsx. Do you want us to try to recover as much as we can? If you trust the source of this workbook, click yes."